# back-end/src/main/java/com/example/employeemanager/auth/ResetPasswordRequest.java
package com.example.employeemanager.auth;

import jakarta.validation.constraints.*;
import lombok.Builder;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@Builder
public class ResetPasswordRequest {
    @Email(message = "Email is not well formatted")
    @NotEmpty(message = "Email is mandatory")
    @NotNull(message = "Email is mandatory")
    @Pattern(regexp = ".+@.+\\..+", message = "Email must be well formatted and contain a domain")
    private String email;

    @NotEmpty(message = "Password is mandatory")
    @NotNull(message = "Password is mandatory")
    @Size(min = 8, message = "Password must be 8 characters long minimum")
    private String newPassword;
}